datacenter-ws/requirements.txt
asgiref==3.11.1
attrs==25.4.0
diff-match-patch==20241021
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
django-import-export==4.4.0
django-reversion==6.1.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysqlclient==2.2.8
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
python-decouple==3.8
pytest==9.0.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tablib==3.9.0
uritemplate==4.2.0
